Category updated fixed and package update fixed

+++ services/category.service.js
@@ -1,25 +1,39 @@
 import categoryRepository from "../repositories/category.repository.js";
+import fileUploadService from "./fileUpload.service.js";
 
 const createCategory = async (data) => {
-
   const newCategory = await categoryRepository.createCategory(data);
   return newCategory;
 };
 
 const updateCategory = async (id, data) => {
-  try {
-    return await categoryRepository.updateCategory(id, data);
-  } catch (error) {
-    throw new Error(`Error in CategoryService updateCategory: ${error}`);
+  //if new image have been added
+  if (data.image_url) {
+    const existCategory = await categoryRepository.getCategoryById(id, null);
+    if (!existCategory) return false;
+
+    var oldImagePath = existCategory.image_url;
   }
+
+  const updatedCategory = await categoryRepository.updateCategory(id, data);
+
+  if (oldImagePath) {
+    await fileUploadService.removeFile(oldImagePath);
+  }
+  return updatedCategory;
 };
 
 const deleteCategory = async (id) => {
-  try {
-    return await categoryRepository.deleteCategory(id);
-  } catch (error) {
-    throw new Error(`Error in CategoryService deleteCategory: ${error}`);
+  const existCategory = await categoryRepository.getCategoryById(id, null);
+  if (!existCategory) return false;
+
+  var oldImagePath = existCategory.image_url;
+  const deletedCategory = await categoryRepository.deleteCategory(id);
+
+  if (oldImagePath) {
+    await fileUploadService.removeFile(oldImagePath);
   }
+  return deletedCategory;
 };
 
 const getCategories = async (tourType, isAdmin) => {
